+++ scoringrules/_crps.py
@@ -67,210 +67,6 @@ def crps_ensemble(
     return crps.ensemble(forecasts, observations, estimator, backend=backend)
 
 
-def twcrps_ensemble(
-    forecasts: "Array",
-    observations: "ArrayLike",
-    v_func: tp.Callable[["ArrayLike"], "ArrayLike"],
-    /,
-    axis: int = -1,
-    *,
-    estimator: str = "pwm",
-    sorted_ensemble: bool = False,
-    backend: "Backend" = None,
-) -> "Array":
-    r"""Estimate the Threshold-Weighted Continuous Ranked Probability Score (twCRPS) for a finite ensemble.
-
-    Computation is performed using the ensemble representation of the twCRPS in
-    [Allen et al. (2022)](https://arxiv.org/abs/2202.12732):
-
-    $$ \mathrm{twCRPS}(F_{ens}, y) = \frac{1}{M} \sum_{m = 1}^{M} |v(x_{m}) - v(y)| - \frac{1}{2 M^{2}} \sum_{m = 1}^{M} \sum_{j = 1}^{M} |v(x_{m}) - v(x_{j})|,$$
-
-    where $F_{ens}(x) = \sum_{m=1}^{M} 1 \{ x_{m} \leq x \}/M$ is the empirical
-    distribution function associated with an ensemble forecast $x_{1}, \dots, x_{M}$ with
-    $M$ members, and $v$ is the chaining function used to target particular outcomes.
-
-    Parameters
-    ----------
-    forecasts: ArrayLike
-        The predicted forecast ensemble, where the ensemble dimension is by default
-        represented by the last axis.
-    observations: ArrayLike
-        The observed values.
-    v_func: tp.Callable
-        Chaining function used to emphasise particular outcomes. For example, a function that
-        only considers values above a certain threshold $t$ by projecting forecasts and observations
-        to $[t, \inf)$.
-    axis: int
-        The axis corresponding to the ensemble. Default is the last axis.
-    backend: str
-        The name of the backend used for computations. Defaults to 'numba' if available, else 'numpy'.
-
-    Returns
-    -------
-    twcrps: ArrayLike
-        The twCRPS between the forecast ensemble and obs for the chosen chaining function.
-
-    Examples
-    --------
-    >>> from scoringrules import crps
-    >>> twcrps.ensemble(pred, obs)
-    """
-    forecasts, observations = map(v_func, (forecasts, observations))
-    return crps_ensemble(
-        forecasts,
-        observations,
-        axis=axis,
-        sorted_ensemble=sorted_ensemble,
-        estimator=estimator,
-        backend=backend,
-    )
-
-
-def owcrps_ensemble(
-    forecasts: "Array",
-    observations: "ArrayLike",
-    w_func: tp.Callable[["ArrayLike"], "ArrayLike"],
-    /,
-    axis: int = -1,
-    *,
-    estimator: tp.Literal["nrg"] = "nrg",
-    backend: "Backend" = None,
-) -> "Array":
-    r"""Estimate the Outcome-Weighted Continuous Ranked Probability Score (owCRPS) for a finite ensemble.
-
-    Computation is performed using the ensemble representation of the owCRPS in
-    [Allen et al. (2022)](https://arxiv.org/abs/2202.12732):
-
-    $$ \mathrm{owCRPS}(F_{ens}, y) = \frac{1}{M \bar{w}} \sum_{m = 1}^{M} |x_{m} - y|w(x_{m})w(y) - \frac{1}{2 M^{2} \bar{w}^{2}} \sum_{m = 1}^{M} \sum_{j = 1}^{M} |x_{m} - x_{j}|w(x_{m})w(x_{j})w(y),$$
-
-    where $F_{ens}(x) = \sum_{m=1}^{M} 1\{ x_{m} \leq x \}/M$ is the empirical
-    distribution function associated with an ensemble forecast $x_{1}, \dots, x_{M}$ with
-    $M$ members, $w$ is the chosen weight function, and $\bar{w} = \sum_{m=1}^{M}w(x_{m})/M$.
-
-    Parameters
-    ----------
-    forecasts: ArrayLike
-        The predicted forecast ensemble, where the ensemble dimension is by default
-        represented by the last axis.
-    observations: ArrayLike
-        The observed values.
-    w_func: tp.Callable
-        Weight function used to emphasise particular outcomes.
-    axis: int
-        The axis corresponding to the ensemble. Default is the last axis.
-    backend: str
-        The name of the backend used for computations. Defaults to 'numba' if available, else 'numpy'.
-
-    Returns
-    -------
-    owcrps: ArrayLike
-        The owCRPS between the forecast ensemble and obs for the chosen weight function.
-
-    Examples
-    --------
-    >>> from scoringrules import crps
-    >>> owcrps.ensemble(pred, obs)
-    """
-    B = backends.active if backend is None else backends[backend]
-
-    if estimator != "nrg":
-        raise ValueError(
-            "Only the energy form of the estimator is available "
-            "for the outcome-weighted CRPS."
-        )
-    if axis != -1:
-        forecasts = B.moveaxis(forecasts, axis, -1)
-
-    fcts_weights, obs_weights = map(w_func, (forecasts, observations))
-
-    if backend == "numba":
-        return crps.estimator_gufuncs["ow" + estimator](
-            forecasts, observations, fcts_weights, obs_weights
-        )
-
-    forecasts, observations, fcts_weights, obs_weights = map(
-        B.asarray, (forecasts, observations, fcts_weights, obs_weights)
-    )
-    return crps.ow_ensemble(
-        forecasts, observations, fcts_weights, obs_weights, backend=backend
-    )
-
-
-def vrcrps_ensemble(
-    forecasts: "Array",
-    observations: "ArrayLike",
-    w_func: tp.Callable[["ArrayLike"], "ArrayLike"],
-    /,
-    axis: int = -1,
-    *,
-    estimator: tp.Literal["nrg"] = "nrg",
-    backend: "Backend" = None,
-) -> "Array":
-    r"""Estimate the Vertically Re-scaled Continuous Ranked Probability Score (vrCRPS) for a finite ensemble.
-
-    Computation is performed using the ensemble representation of the vrCRPS in
-    [Allen et al. (2022)](https://arxiv.org/abs/2202.12732):
-
-    \[
-    \begin{split}
-        \mathrm{vrCRPS}(F_{ens}, y) = & \frac{1}{M} \sum_{m = 1}^{M} |x_{m} - y|w(x_{m})w(y) - \frac{1}{2 M^{2}} \sum_{m = 1}^{M} \sum_{j = 1}^{M} |x_{m} - x_{j}|w(x_{m})w(x_{j}) \\
-            & + \left( \frac{1}{M} \sum_{m = 1}^{M} |x_{m}| w(x_{m}) - |y| w(y) \right) \left( \frac{1}{M} \sum_{m = 1}^{M} w(x_{m}) - w(y) \right),
-    \end{split}
-    \]
-
-    where $F_{ens}(x) = \sum_{m=1}^{M} 1 \{ x_{m} \leq x \}/M$ is the empirical
-    distribution function associated with an ensemble forecast $x_{1}, \dots, x_{M}$ with
-    $M$ members, $w$ is the chosen weight function, and $\bar{w} = \sum_{m=1}^{M}w(x_{m})/M$.
-
-    Parameters
-    ----------
-    forecasts: ArrayLike
-        The predicted forecast ensemble, where the ensemble dimension is by default
-        represented by the last axis.
-    observations: ArrayLike
-        The observed values.
-    w_func: tp.Callable
-        Weight function used to emphasise particular outcomes.
-    axis: int
-        The axis corresponding to the ensemble. Default is the last axis.
-    backend: str
-        The name of the backend used for computations. Defaults to 'numba' if available, else 'numpy'.
-
-    Returns
-    -------
-    vrcrps: ArrayLike
-        The vrCRPS between the forecast ensemble and obs for the chosen weight function.
-
-    Examples
-    --------
-    >>> from scoringrules import crps
-    >>> vrcrps.ensemble(pred, obs)
-    """
-    B = backends.active if backend is None else backends[backend]
-
-    if estimator != "nrg":
-        raise ValueError(
-            "Only the energy form of the estimator is available "
-            "for the outcome-weighted CRPS."
-        )
-    if axis != -1:
-        forecasts = B.moveaxis(forecasts, axis, -1)
-
-    fcts_weights, obs_weights = map(w_func, (forecasts, observations))
-
-    if backend == "numba":
-        return crps.estimator_gufuncs["vr" + estimator](
-            forecasts, observations, fcts_weights, obs_weights
-        )
-
-    forecasts, observations, fcts_weights, obs_weights = map(
-        B.asarray, (forecasts, observations, fcts_weights, obs_weights)
-    )
-    return crps.vr_ensemble(
-        forecasts, observations, fcts_weights, obs_weights, backend=backend
-    )
-
-
 def crps_beta(
     shape1: "ArrayLike",
     shape2: "ArrayLike",
@@ -1571,9 +1367,6 @@ def crps_tpnorm(
 
 __all__ = [
     "crps_ensemble",
-    "twcrps_ensemble",
-    "owcrps_ensemble",
-    "vrcrps_ensemble",
     "crps_beta",
     "crps_binomial",
     "crps_exponential",
